back/gapsi_ecommerce/requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
flake8==7.0.0
marshmallow==3.20.2
mccabe==0.7.0
packaging==23.2
pycodestyle==2.11.1
pyflakes==3.2.0
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4
